Remove run.bat and added install.py ! Remove virtual environnement
For the new installation and for the new use of PyCub you will just have to launch the script install.py (you make a cd of the folder and you write install.py) the script will make you save 4 lines of commands and will create a shortcut on your desktop (windows or mac) to start the terminal of PyCub!
