+++ README.md
@@ -1,7 +1,3 @@
-
 # pycub
 A transpiler of simplified source code (the pycub language dedicated to coding on Minecraft) into Java code, the project was inspired by the organization of SkriptLang
 Official Discord: https://discord.gg/hm8VXyVx5u
-
-# Vidéo exemple for how to manage and create your first project :p (Alpha 5.0)
-https://user-images.githubusercontent.com/24825008/118866236-133fcc80-b8e2-11eb-85a1-4a60a34f5437.mp4
