--- projects/fiffa/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sun May 23 19:07:56 CEST 2021
-gradle.version=5.6.1
