--- projects/test/src/main/resources/plugin.yml
@@ -1,5 +1,0 @@
-name: test
-author: PyCubTranspilator
-main: fr.exemple.author
-version: 0.1
-commands:
