--- run.bat
@@ -1,2 +1,0 @@
-env\Scripts\activate.bat
-python main.py
