--- scripts/test.pycub
@@ -1,14 +1,0 @@
-start file: startfile
-
-on player join:
-    broadcast(player name + "testsetseest")
-    stop
-
-change file: omgg
-
-on player join:
-    broadcast(player name)
-    stop
-
-change file: testsetseest
-
